### 2025-01-09 #170 @editorialbot
[PRE REVIEW]: DMRGenie: Graphical interface for entanglement renormalization
**Submitting author:** @bakerte (<a href="http://orcid.org/0000-0002-3142-0767">Thomas Baker</a>)
**Repository:** https://github.com/bakerte/DMRGenie.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
**Managing EiC:** Ludovic Räss

Thanks for submitting your paper to JCON @bakerte. **Currently, there isn't a JCON editor assigned** to your paper.
