### 2025-01-09 2c37a49
fixes to DMRGenie JS packages and changed TensEZ to DMRGenie
