+++ paper/paper.tex
@@ -207,7 +207,7 @@ DMRGenie can be run locally and we show how to do this here.
 
 \subsection{Setup}
 
-Navigate to the `tensez` directory once the repository has been cloned and then into the `TensEZ` folder within it.
+Navigate to the `tensez` directory once the repository has been cloned and then into the `DMRGenie` folder within it.
 
 Install the necessary packages with instantiate.
 
@@ -218,7 +218,7 @@ pkg> instantiate
 
 \subsection{Running the App}
 
-The app must be activated by running the following from the `TensEZ` directory
+The app must be activated by running the following from the `DMRGenie` directory
 
 \begin{lstlisting}[language = Julia]
 julia> ] 
@@ -355,7 +355,7 @@ When an operation is done, it passes the name of the operation done and the Java
 \end{itemize}
 
 
-%Make sure that the path ends with `tensez/TensEZ`, if not then run `{\tt cd("<path\_to\_repo>/tensez/TensEZ")}` from the Julia CLI.
+%Make sure that the path ends with `tensez/DMRGenie`, if not then run `{\tt cd("<path\_to\_repo>/tensez/DMRGenie")}` from the Julia CLI.
 %
 %Then exit the Pkg prompt with *backspace*.
 %
