+++ paper/paper.yml
@@ -16,6 +16,7 @@ authors:
     orcid: 0009-0009-0617-9500
     affiliation: 2
   - name: Sarah Huber
+    orcid: 0000-0002-3200-0262
     affiliation: 2
   - name: Thomas E. Baker
     orcid: 0000-0002-3142-0767
