+++ paper/paper.yml
@@ -11,6 +11,7 @@ authors:
   - name: Kiana Gallagher
     affiliation: 1
   - name: Aaron Dayton
+    orcid: 0009-0003-7573-6143
     affiliation: 1
   - name: Drew Leske
     orcid: 0009-0009-0617-9500
