+++ Project.toml
@@ -31,9 +31,9 @@ UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 Genie = "=5.18"
 HTTP = "=1.9"
 
-julia = "1.11.0"
-DMRJtensor = "=0.15.0"
-TensorPACK = "=0.9.0"
+julia = "=1.9.3"
+DMRJtensor = "=0.11.4"
+TensorPACK = "=0.5.2"
 LinearAlgebra = "1.11.0"
 Distributed = "1.11.0"
 Printf = "1.11.0"
