### 2025-01-13 #170 @luraess
Dear authors, thank you for submitting your work to the JuliaCon Proceedings. @bakerte could you please try to check and fix the invalid an missing DOIs and make sure none exist for the skipped once (see https://github.com/JuliaCon/proceedings-review/issues/170#issuecomment-2580585471). Also, there seems to be an issue with PDF generation (see https://github.com/JuliaCon/proceedings-review/issues/170#issuecomment-2580584653). In the meantime, I will be looking for reviewers.
