### 2025-01-15 0367638
Gitignore and stuff
