+++ DevelopersGuide/DevelopersGuide.tex
@@ -1,0 +1,195 @@
+\documentclass{article}
+\usepackage[utf8]{inputenc}
+\usepackage{geometry}
+\usepackage[T1]{fontenc}
+\usepackage{amsfonts}
+\usepackage{graphicx}
+\usepackage{float}
+\usepackage{rotating}
+\usepackage{hyperref}
+\usepackage[sorting=none]{biblatex}
+\usepackage{fancyhdr}
+\usepackage{multicol}
+\usepackage{braket}
+\usepackage{amssymb}
+\usepackage{array}
+\addbibresource{./ref.bib}
+\setlength{\columnsep}{40pt}
+\setlength{\voffset}{0.7cm}
+\setlength{\headsep}{40pt}
+\geometry{letterpaper, portrait, margin=2cm}
+\graphicspath{ {./Figures/} }
+
+\usepackage{lmodern}  % for bold teletype font
+\usepackage{amsmath}  % for \hookrightarrow
+\usepackage{xcolor}   % for \textcolor
+\usepackage{listings}
+\lstset{
+  basicstyle=\ttfamily,
+  columns=fullflexible,
+  frame=single,
+  breaklines=true,
+  postbreak=\mbox{\textcolor{red}{$\hookrightarrow$}\space},
+  showstringspaces=false
+}
+
+\usepackage[export]{adjustbox}
+
+\usepackage[center]{titlesec}
+\renewcommand{\thesection}{\Roman{section}} 
+\renewcommand{\thesubsection}{\Alph{subsection}}
+\renewcommand{\thesubsubsection}{\Alph{subsection}.\textit{\arabic{subsubsection}}}
+
+% Title page
+\title{DMRGenie Developer's Guide}
+\author{Aaron Dayton}
+\date{\today}
+
+% Header and footer
+\pagestyle{fancy}
+\fancyhead{}
+\fancyfoot{}
+\begin{document}
+
+\maketitle
+\thispagestyle{fancy}
+\clearpage
+\tableofcontents
+\thispagestyle{fancy}
+
+\clearpage
+% Begin page numbers
+\fancyhead[C]{\thepage}
+\pagenumbering{arabic}
+
+\section{ABSTRACT}
+
+\linebreak
+\begin{multicols}{2}
+\section{INTRODUCTION}
+
+This document covers how to add/modify features in the DMRGenie webapp. DMRGenie is built on a Model View Controller (MVC) infrastructure. For information on the implementation of the MVC framework with basic functionality, the package decisions, or how to run the app, see the \lstinline|Readme.md| file. Some information is common between these documents.
+
+Many instructions can be found in the Genie documentation \href{https://genieframework.github.io/Genie.jl/dev}{here}. For example, to learn how to create a new MVC group read \href{https://genieframework.github.io/Genie.jl/dev/tutorials/4-1--Developing_MVC_Web_Apps.html}{this} page.
+
+\section{File structure}
+
+The main folder of concern for editing the application is \lstinline|DMRGenie/app/resources|. Within this folder, all the models, views, and controllers are sorted into distinct folders. As of May 16, 2024, the only group included is the \lstinline|tensornetwork| group. A supporting folder is \lstinline|DMRGenie/public|. The file \lstinline|DMRGenie/routes.jl| handles the exchange of data between the front and back end.
+
+\subsection{The \lstinline|tensornetwork| MVC group}
+
+Within the resources folder is the \lstinline|tensornetworks| directory, which contains \lstinline|TensorNetworks.jl|, \lstinline|TensorNetworksController.jl|, \lstinline|views|, and \lstinline|TensorNetworksValidator.jl|.
+
[…]
