+++ DMRGenie/run.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+julia -i run.jl
