+++ README.md
@@ -12,66 +12,6 @@ Please note that **Julia version 1.9.x** is required for local use.
 
 Questions? Email [bakerte@uvic.ca](bakerte@uvic.ca)
 
-## Using the App
-
-There are four main sections of the application as of Aug 10, 2023: the Model input, the Method input, the Measurement output, and the reactive display which shows your current tensor network.
-
-### Model
-
-The *Model* section first takes information about whether or not a default model is to be used, if no default model is used then the user is required to create a custom Hamiltonian.
-It then takes information about whether or not correlations should be output, and if true, it asks what the desired operators for the correlation matrices and functions are.
-Then it takes symmetry information. If symmetry is checked, then DMRG and the desired correlations will be computed for the symmetry data specified.
-The symmetry data taken includes the name of the symmetry, which can be any string; the symmetry operators to be applied, which there can be multiple of; whether or not the symmetries are uniform, and the physical indices.
-
-If the symmetries are specified to be non-uniform, then multiple physical indices can be specified with the range of which dimensions they will be applied over.
-**NOTE:** Currently there are only 2 physical dimensions allowed, so the range can only act over 1-2. **Control over the number of physical dimensions is to be allowed in a future update.**
-
-### Method
-
-In the *Method* section the graph type can be specified (currently this is always assumed to be an MPO on the back-end, regardless of the input, **functionality must be added**). Then the number of tensors in the network is specified, increasing this greatly increases the time to run calculations, especially if large correlation functions are to be calculated. The system can be generated randomly, or input by the user in .txt or .csv format, **functionality must be added for systems input by the user**. Then the tensor network method to be run on the system is specified, currently this is not used and is subject to change depending on the requirements of the application.
-
-### Measurements
-
-In the *Measurements* section no input is taken, but after the form is submitted and results are calculated, those results are output to the user. Currently, DMRG with and without symmetries can be output, as well as correlation matrices and correlation functions with and without symmetries. All DMRG calculations are given as a number, and all correlation calculations are given as downloadable text files. So there are currently 6 outputs which can appear to the user.
-
-### Tensor Network Display
-
-The *Display* shows the user what the tensor network to be used in calculations looks like.
-
-## Genie and Model-View-Controller Frameworks
-
-This application uses a Julia package called Genie which allows Julia code to be put on a web server simply.
-Genie is a Model-View-Controller (MVC) framework.
-
-A Model in an MVC framework is basically the application's structure. For this application, the main model is the `TensorNetwork` model, which can be seen in `DMRGenie/app/resources/tensornetworks/TensorNetworks.jl`. There may be more models added later depending on what is required.
-
-A View in an MVC framework is what is displayed to the user: the front-end. An application can have many views, one for each page.
-The main view of DMRGenie is `DMRGenie/app/resources/tensornetworks/views/DMRGenie.jl.html`, a `.jl.html file`. A `.jl.html` file acts as a normal HTML file with the added benefit of being able to run Julia code and have Julia variables passed in.
-This allows data calculated on the back-end to be passed to the front-end and be presented to the user.
-Views are linked to through `routes.jl -> specified controller -> specified view`.
-
-Controllers in an MVC framework are the meat and potatoes of the application, the back-end.
-In DMRGenie, the controllers calculate all the scientific results given to the user such as DMRG and the correlation matrices for a tensor network.
-Julia code is run to obtain data, which is then passed to the view with the `html` function of Genie.
-
-All of this is mediated by `routes.jl`, which directs the traffic of the app and gives any data submitted to the front-end by the user to the back-end.
-
-### Why Genie?
-
[…]
