### 2025-01-15 3bc1817
removed required field from some hidden inputs on alg runner
