### 2025-01-20 #170 @bakerte
Dear Editor,

We have implemented all DOIs that we could identify, including all of those marked with a red X on the message above. Not all references in the "ignored" column could have DOIs assigned, but we can try again if requested.

We have also pushed a new version and created a new branch version for the /paper directory branch (if your system needs to be updated because we indicated the branch, then we note that here).

Lastly, we were able to compile the paper in the directory and now include the pdf generated from the instructions given on the JuliaCon submission site. Both `ruby prep.rb` and `latexmk -bibtex -pdf paper.tex` do not encounter the error, so we could not reproduce it on our machines. Is it possible that there is a discrepancy between our LaTeX distribution and the one you are using? We are using the latest version updated a few weeks go. We are happy to try to solve this problem in another way if it still does not work.

Best regards,
Thomas E. Baker
