### 2025-01-20 2419e6d
Update paper.yml
