+++ paper/paper.yml
@@ -23,13 +23,13 @@ authors:
     orcid: 0000-0002-3142-0767
     affiliation: 1,3,4
 affiliations:
-  - name: Department of Physics \& Astronomy, University of Victoria, Victoria, British Columbia V8P 5C2, Canada
+  - name: "Department of Physics \& Astronomy, University of Victoria, Victoria, British Columbia V8P 5C2, Canada"
     index: 1
-  - name: Research Computing Services, University of Victoria, Victoria, British Columbia V8P 5C2, Canada
+  - name: "Research Computing Services, University of Victoria, Victoria, British Columbia V8P 5C2, Canada"
     index: 2
-  - name: Department of Chemistry, University of Victoria, Victoria, British Columbia V8P 5C2, Canada
+  - name: "Department of Chemistry, University of Victoria, Victoria, British Columbia V8P 5C2, Canada"
     index: 3
-  - name: Centre for Advanced Materials and Related Technologies, University of Victoria, Victoria, British Columbia V8P 5C2, Canada
+  - name: "Centre for Advanced Materials and Related Technology, University of Victoria, Victoria, British Columbia V8P 5C2, Canada"
     index: 4
 date: 31 December 2024
 bibliography: ref.bib
