### 2025-01-20 #170 @bakerte
Highly curious that does not work here. Let me try one more itme with the quotes around the affiliations in the .yml file
