### 2025-01-20 #170 @luraess
Thanks for trying. Note, the bot commands should be first in comments else they don't get executed.
