### 2025-01-20 #170 @luraess
Thanks for your notice. There could be an issue with the `\&` sequence in there yaml file, maybe. Thanks for looking into this further in the coming days!
