### 2025-01-20 #170 @luraess
Could be that spaces are needed here (`1, 2, 3`) https://github.com/bakerte/DMRGenie.jl/blob/fc896f746d7e9fe24eb84fb8c463ddb181013932/paper/paper.yml#L24
