+++ paper/paper.yml
@@ -21,7 +21,7 @@ authors:
     affiliation: 2
   - name: Thomas E. Baker
     orcid: 0000-0002-3142-0767
-    affiliation: 1,3,4
+    affiliation: 1, 3, 4
 affiliations:
   - name: Department of Physics and Astronomy, University of Victoria, Victoria, British Columbia V8P 5C2, Canada
     index: 1
