### 2025-01-31 #170 @luraess
@asinghvi17 and @dpsanders - we have this exciting JuliaCon submission that might be in your domain of interest (GUI, visualisation). Let me know in the next couple of days if you'd be willing to help out with the revision. Thanks
