### 2025-02-06 #170 @bakerte
Hello,

It is tricky. I know of people on the side of the entanglement renormalization concepts, but not so much on GUIs. I will say that the back half of the information is mainly tailored towards anyone with experience with GUIs and not necessarily the physics application.

Chris Raucaukas might know about my work because he was at the same institution as I was when I created this and we have crossed paths a few times.
Laurens Vanderstraeten has a background in tensor networks for these types of problems.
Philippe Corboz is also someone who might be interested in this

I’m not sure how much more information I can post on this public forum. Please let me know if I should think of more names.
