### 2025-03-19 d7ec0e4
Removed limitations on DMRGenie for local users
