### 2025-06-05 #183 @editorialbot
[REVIEW]: DMRGenie: Graphical interface for entanglement renormalization
**Submitting author:** @bakerte (<a href="http://orcid.org/0000-0002-3142-0767">Thomas Baker</a>)
**Repository:** https://github.com/bakerte/DMRGenie.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
