### 2025-06-05 #183 @luraess
@cortogantese @AAyati1, please create your checklist typing: `@editorialbot generate my checklist`. From there, please proceed with the revision as following:
- For issues/suggestions/improvements directly related with the https://github.com/bakerte/DMRGenie.jl package, open an issue there directly referring to this revision PR.
- For anything related to the manuscript itself, feel free to post a report using this PR.
Note that there is no formal need for a "statement of need" section in the paper, but this information should be contained in the first parts of the manuscript.
Thank you for your help, and do not hesitate to reach out if you have any questions
