### 2025-11-12 #183 @bakerte
Hi there,

Yes, I am still monitoring this. I was going to deal with this once I return from leave, but if you have suggestions for how to move this forward, please let me know. I could send an email to the referee if that would help, but I wasn’t sure if that was appropriate.

I could also suggest new referees. What would be best here?

> On Nov 12, 2025, at 09:09, Ludovic Räss ***@***.***> wrote:
>
>
> luraess
>  left a comment
> (JuliaCon/proceedings-review#183)
>  <https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3522866896>
> Can we get this moving forward, or if you feel this submission is no longer a thing, let me know so we can terminate the process. It's now soon half a year without any action nor comment from anyone, reviewers nor authors. I'll set a reminder and if no answers nor actions are undertaken within the next 2 weeks, I'll consider this submission as stall and will see to close it @bakerte <https://github.com/bakerte> @AAyati1 <https://github.com/AAyati1> @cortogantese <https://github.com/cortogantese>
