### 2025-11-12 #183 @AAyati1
Hello,

I am so sorry for this late reply. This is an old email address that I
haven’t checked in a while. I’d completely forgotten about this and hadn’t
seen these remimders. I apologize and will write my comments and review by
tomorrow.

Once again I apologize for this mistake.

Regards,
Amir

On Wed, Nov 12, 2025 at 1:29 PM Thomas E. Baker ***@***.***>
wrote:

> *bakerte* left a comment (JuliaCon/proceedings-review#183)
> <https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3523078375>
> Hi there,
>
> Yes, I am still monitoring this. I was going to deal with this once I
> return from leave, but if you have suggestions for how to move this
> forward, please let me know. I could send an email to the referee if that
> would help, but I wasn’t sure if that was appropriate.
>
> I could also suggest new referees. What would be best here?
>
> > On Nov 12, 2025, at 09:09, Ludovic Räss ***@***.***> wrote:
> >
> >
> > luraess
> > left a comment
> > (JuliaCon/proceedings-review#183)
> > <
> https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3522866896>
>
> > Can we get this moving forward, or if you feel this submission is no
> longer a thing, let me know so we can terminate the process. It's now soon
> half a year without any action nor comment from anyone, reviewers nor
> authors. I'll set a reminder and if no answers nor actions are undertaken
> within the next 2 weeks, I'll consider this submission as stall and will
> see to close it @bakerte <https://github.com/bakerte> @AAyati1 <
> https://github.com/AAyati1> @cortogantese <https://github.com/cortogantese>
>
> > —
> > Reply to this email directly, view it on GitHub <
> https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3522866896>,
> or unsubscribe <
> https://github.com/notifications/unsubscribe-auth/AB3JZYU32V7BHVXOBFWV3UL34NSU3AVCNFSM6AAAAAB6WOXWXCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTKMRSHA3DMOBZGY>.
>
> > You are receiving this because you were mentioned.
> >
>
