### 2025-11-26 42173a8
Update paper.tex
