### 2025-11-26 48ec403
Update paper.tex
