### 2025-12-23 #183 @bakerte
Dear Editor,

First, happy holidays and a happy new year to you and your’s. I wanted to check signals on the status of this paper. Let me just say that I am still interested in publishing with JOSS for the JuliaCon proceedings.

Since the paper has taken about a year to get reviews, I thought some productive dialogue would help us to resolve this bottleneck and clear this from our mutual agendas. I think the paper is fairly straightforward. The online tool is a faithful deployment of the software that we developed and that was of interest at the JuliaCon conference. In my view, a review of this paper would acknowledge that the tool works and allows for the running/development of future algorithms using tensor networks, particularly in the case of quantum physics where entanglement and renormalization are proven pathways to do so.

I thought the few remaining objectives in the text were to communicate how we made the tool to the Julia community (we heard at the conference that the tool didn’t even look like Julia, which we took as a compliment), explain the network of packages we have developed, and also to give a bare introduction to the field with references to previous, longer treatments of the subject. I realise that finding referees for this case may be a challenge since the paper is interdisciplinary along the lines I have just mentioned. My referee suggestions were just to get a perspective from quantum chemistry, mainly, to see if they had suggestions for the advertisement of this tool to them. Perhaps I asked too much of the referees despite trying to set this up to be easy for them.

The GUI we developed is still running and will undergo some improvements and code updates over the next year. We had a three year development plan for this one and I thought that talking to you about how to forward this might give a boost to the students who developed this tool.

I’m not sure what to do, exactly, so let’s just discuss what we can do. I will offer a few options. I suppose I could suggest a few more referees, although I see other things that we could do. What do you suggest at this point? Please feel free to leave this to the new year if that is easier for you. I continue to appreciate your stewardship of this paper during this process.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
