### 2026-01-05 #183 @luraess
Dear Thomas Baker,

Thank you for your message and suggestions. I would still like to see your paper published in the Proceedings so let's try to find at least one other reviewer (possibly two) that would be willing to take a look at the draft. It would be helpful to get your input about referees willing to do so from your field. Else, I can also screen on my side for more general Julia feedback.

Thanks for your help and willingness to push this forward, and happy 2026.

Best, Ludovic Räss
