### 2026-01-23 #183 @bakerte
Dear Editor,

Here is a list of a few names. I can provide more if there are no takers or if it drags out too long.

Alberto Nocera
***@***.***

Adrian Feiguin
***@***.***

Goncalo Catarina
***@***.***

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
