### 2026-01-28 #183 @bakerte
Hello,

Alberto Nocera: @alberto242
Adrian Feiguin: @afeiguin

There appear to be a few GitHub handles related to the third name. Perhaps we can just try the first two and see if we get anywhere.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
