### 2026-01-28 #183 @luraess
Dear @afeiguin,

I am reaching out to ask whether you'd be willing to support the review process of this JuliaCon Proceedings submission?

Please have look to the statement by the author https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3687782702 to get a better idea about the submission.

The paper can be access here https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-2946227087.

I am looking forward to hearing back from you!
