### 2026-01-28 #183 @afeiguin
I’d be happy to do it. However, I’d appreciate a little information about the review process, any prior correspondence, and how to submit my recommendation.

Best,
<ADrIAN>

> On Jan 28, 2026, at 12:00 PM, Ludovic Räss ***@***.***> wrote:
>
>
> luraess
>  left a comment
> (JuliaCon/proceedings-review#183)
>  <https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3812574391>
> Dear @afeiguin <https://github.com/afeiguin>,
>
> I am reaching out to ask whether you'd be willing to support the review process of this JuliaCon Proceedings submission?
>
> Please have look to the statement by the author #183 (comment) <https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3687782702> to get a better idea about the submission.
>
> The paper can be access here #183 (comment) <https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-2946227087>.
>
> I am looking forward to hearing back from you!
>
