### 2026-01-28 #183 @luraess
> I’d appreciate a little information about the review process, any prior correspondence, and how to submit my recommendation.

Thank you for accepting to review this work. The review is GitHub-based opening issues with comments, issues, fixes, and suggestions to both the code and paper repo. More details can be found hereafter, and do not hesitate to ping me if you have further questions.

**Reviewers and authors**:

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/bakerte/DMRGenie.jl) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

## Reviewer instructions & questions

@afeiguin, your review will be checklist based. You will have a separate checklist that you should update when carrying out your review. First of all you need to run this command in a separate comment to create the checklist:

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let [@luraess](https://github.com/luraess) know.
