### 2026-01-28 #183 @afeiguin
I am seriously busy and honestly I don’t have time to learn all the minutiae of your review system. I can read the paper, look at the software and answer any questions but don’t ask me to go over all these steps and follow these obscure instructions that I don’t understand Sent from my iPhoneOn Jan 28, 2026, at 16:11, Ludovic Räss ***@***.***> wrote:﻿luraess left a comment (JuliaCon/proceedings-review#183)

I’d appreciate a little information about the review process, any prior correspondence, and how to submit my recommendation.

Thank you for accepting to review this work. The review is GitHub-based opening issues with comments, issues, fixes, and suggestions to both the code and paper repo. More details can be found hereafter, and do not hesitate to ping me if you have further questions.
Reviewers and authors:
Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
Reviewer instructions & questions
@afeiguin, your review will be checklist based. You will have a separate checklist that you should update when carrying out your review. First of all you need to run this command in a separate comment to create the checklist:

Checklists
@afeiguin, please create your checklist typing: @editorialbot generate my checklist
