### 2026-01-29 #183 @luraess
Here is a summary of the actions to help with the review process.

If the GitHub-based workflow via issues is too unfamiliar to you, feel free to post any comments or issues as a comment in this conversation or upload a file in there as well. The checklist below will guide you.

## Download the proof

[Download article proof](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00183/jcon.00183/10.21105.jcon.00183.pdf)

## Review checklist

Use the following checklist to guide your revision process. The software is available at https://github.com/bakerte/DMRGenie.jl

- [ ]  I confirm that I have read the [JuliaCon conflict of interest policy](https://juliacon.github.io/proceedings-guide/reviewer/#conflict_of_interest) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JCon for the purpose of this review.

- [ ]  **Repository:** Is the source code for this software available at the https://github.com/bakerte/DMRGenie.jl?
- [ ]  **Contribution and authorship:** Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [ ]  **Authors:** Does the `paper.pdf` file include a list of authors with their affiliations?
- [ ]  **A statement of need:** Does the paper clearly explain the problem it addresses, its importance to the broader community, the intended audience, and how it connects to existing work in the field?
