### 2026-03-26 #183 @bakerte
Dear Editor,

I think we may need to move on to a new referee. I was going to suggest Prof. Matthew Amy ***@***.******@***.***>) from SFU. We have met each other once or twice. I believe in his group he would have someone that could navigate both the paper and the git interface for reviews.

I refreshed the “paper” branch since I think there were changes on the main file that we had internally that didn’t get pushed there. Let me know if I should do anything else. If we don’t hear back in a few weeks, I will suggest a new reviewer.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
