### 2026-03-27 #183 @luraess
Hi @meamy, I am reaching out as we are currently looking for reviewer to provide feedback on the JuliaCon Proceedings submission by [Thomas Baker](http://orcid.org/0000-0002-3142-0767): [👉📄 Download article proof](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00183/jcon.00183/10.21105.jcon.00183.pdf). The authors suggested your name as possible reviewer. If willing to help out, please let me know and I could add you to the reviewer and generate your checklist. A short overview of the GitHub based review process is available e.g. here https://github.com/JuliaCon/proceedings-review/issues/183#issuecomment-3813952160.
