### 2026-04-15 #183 @bakerte
Dear Editor,

We are discussing internally amongst ourselves about who we can recommend here. I fear my network may not be sufficient as we have seen, so we may go with someone more on the software side.

I wanted to ask for your advice on this. I expected the referees to look at the online tool that we deployed, see that it works and looks accessible, and then read the paper. Potentially some small comments will come out of that, but I thought this was an easy acceptance. As we saw from one of the referees, they were uneasy about the online interface. I suppose an open interview is also novel for people in my area. Have you noticed anything that we should take into account here? I’m perplexed why I have not been able to identify a willing reviewer so far.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
