### 2026-04-21 #183 @luraess
Maybe folks suggested are not familiar with the JOSS-style papers, submissions, and that actually the software sits at the core of the submission. Maybe you have an enthusiastic user of the software that would be willing to go over the work. If so, you could let me know. Else, I will try to find someone willing to review the functionality of the proposed work but with limited knowledge about the underlying domain science.
