### 2026-04-22 #183 @bakerte
Thank you for sharing your view on this. We had a similar discussion internally and will get back to you with a list. I have circulated my colleagues on this. Let me apologise as well, the work of our group has exploded out of traditional quantum information and I should have understood when suggesting previous referees.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
