### 2026-05-06 #183 @bakerte
Dear Editor,

We were not able to identify anyone still in an academic environment, nor anyone suitable. Our links to computer science groups has atrophied, unfortunately. I could reach out to colleagues in Computer Science, but if you have someone in mind, we would welcome the opportunity to talk to them.

We would also be willing to adjust the manuscript towards a specific audience if it would help in unlocking a pool of referees. Thank you for helping us with this.

Best regards,
Thomas Baker
_________________________________
Thomas E. Baker
Canada Research Chair in Quantum Computing for Modelling of Molecules and Materials (website<https://sites.google.com/view/bakerte>)
Member of Quantum BC<https://quantum-bc.ca/>
Department of Physics & Astronomy
Department of Chemistry
University of Victoria, British Columbia, Canada
Territory acknowledgement<https://www.uvic.ca/services/indigenous/facultystaff/territory-acknowledgment/index.php>
